+++ src/QMCDrivers/CMakeLists.txt
@@ -36,7 +36,7 @@ SET(QMCDRIVERS
   WaveFunctionTester.cpp
   WalkerControlBase.cpp
   CloneManager.cpp
-  MoveContext.cpp
+  ContextForSteps.cpp
   QMCUpdateBase.cpp
   GreenFunctionModifiers/DriftModifierBuilder.cpp
   GreenFunctionModifiers/DriftModifierUNR.cpp
